Port scanner
The following code allows the user to select a host using a domain name, a single IP address, an IP address with a subnet mask, or a list of IPs from a text file. The user can choose to scan a single or range of ports. The user can request to run a UDP scan; otherwise, the default is TCP. The verbose option allows users to see which port is currently being scanned; otherwise, they see only open ports. Once the scan is completed, the user sees how long it took. They also have the option of viewing the results in a PDF report.
Example commands: portscanner.py -d localhost -rng 440-445 -0f coolreport -v; portscanner.py -sub 192.168.1.0/24 -prt 443;
-d, --domain (Enter a host to scan. Ex. www.google.com)
-IP, --IPaddress (Enter an IP address to scan. Ex. 127.0.0.1)
-sub, --subnet (Enter IP address with subnet mask to scan.)
-if --ifile (Enter the name of a text file with IP addresses to scan. Addresses must be separated by a single space. Ex. list.txt)
-prt --port (Enter a single port number. Ex. 443)
-rng, --range (Enter a range of ports. Ex. 440-445)
-of, --ofile (Enter a name for the pdf report. Ex. Myreport)
--UDP (Uses UDP if true)
-v, --verbose (Displays ports while scanning. Otherwise, displays only open ports)